+++ src/Captcha/RecaptchaV2Provider.php
@@ -4,7 +4,7 @@
  *
  * @author    Arte e Informatica <helpdesk@tecnoacquisti.com>
  * @copyright 2009-2026 Arte e Informatica
- * @license   One Paid Licence By WebSite Using This Module. No Rent. No Sell. No Share.
+ * @license   MIT License
  */
 
 namespace TecSpamGuard\Captcha;